Re-new code logical User Module

+++ Backend/Gigantic-Web-Parent/Gigantic-Management/src/main/java/com/gigantic/user/repository/RoleRepository.java
@@ -1,9 +1,9 @@
 package com.gigantic.user.repository;
 
-import com.gigantic.entity.Role;
+import com.gigantic.user.model.Role;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RoleRepository extends CrudRepository<Role,Long> {
+public interface RoleRepository extends CrudRepository<Role,Integer> {
 }
